+++ pool/app/i18n/entity.ml
@@ -12,6 +12,7 @@ module Key = struct
       | GreetingsText [@name "greetings_text"] [@printer print "greetings_text"]
       | PasswordPolicyText [@name "password_policy_text"]
           [@printer print "password_policy_text"]
+      | PrivacyPolicy [@name "privacy_policy"] [@printer print "privacy_policy"]
       | WelcomeText [@name "welcome_text"] [@printer print "welcome_text"]
     [@@deriving enum, eq, ord, sexp_of, show { with_path = false }, yojson]
   end
@@ -20,7 +21,7 @@ module Key = struct
   include Core
 
   let is_rich_text = function
-    | CreditsText | GreetingsText | WelcomeText -> true
+    | CreditsText | GreetingsText | PrivacyPolicy | WelcomeText -> true
     | PasswordPolicyText -> false
   ;;
 end