+++ pool/app/i18n/i18n.mli
@@ -3,6 +3,7 @@ module Key : sig
     | CreditsText
     | GreetingsText
     | PasswordPolicyText
+    | PrivacyPolicy
     | WelcomeText
 
   val create : string -> (t, Pool_common.Message.error) result
@@ -54,13 +55,34 @@ val show_event : event -> string
 val handle_event : Pool_database.Label.t -> event -> unit Lwt.t
 val find : Pool_database.Label.t -> Pool_common.Id.t -> t Lwt.t
 
+val find_with_default_content
+  :  Pool_database.Label.t
+  -> Pool_common.Id.t
+  -> t Lwt.t
+
 val find_by_key
   :  Pool_database.Label.t
   -> Key.t
   -> Pool_common.Language.t
   -> t Lwt.t
 
+val find_by_key_opt
+  :  Pool_database.Label.t
+  -> Key.t
+  -> Pool_common.Language.t
+  -> t option Lwt.t
+
 val find_all : Pool_database.Label.t -> unit -> t list Lwt.t
+
+module I18nPageCache : sig
+  val clear : unit -> unit
+end
+
+val i18n_is_set
+  :  Pool_database.Label.t
+  -> Pool_common.Language.t
+  -> Key.t
+  -> bool Lwt.t
 
 module Guard : sig
   module Target : sig