+++ pool/app/system_event/system_event.mli
@@ -3,6 +3,7 @@ module Id : module type of Pool_common.Id
 module Job : sig
   type t =
     | GuardianCacheCleared
+    | I18nPageUpdated
     | SmtpAccountUpdated
     | TenantDatabaseAdded of Pool_database.Label.t
     | TenantDatabaseUpdated of Pool_database.Label.t
@@ -34,7 +35,7 @@ module EventLog : sig
   module ServiceIdentifier : sig
     include Pool_common.Model.StringSig
 
-    val get : unit -> t
+    val get : ?identifier:string -> unit -> t
   end
 
   module Status : sig
@@ -65,9 +66,10 @@ val pp_event : Format.formatter -> event -> unit
 val show_event : event -> string
 val find_pending : EventLog.ServiceIdentifier.t -> t list Lwt.t
 val handle_event : event -> unit Lwt.t
-val handle_system_event : t -> unit Lwt.t
+val handle_system_event : ?identifier:string -> t -> unit Lwt.t
 
 module Service : sig
-  val run : unit -> unit Lwt.t
-  val register : unit -> Sihl.Container.Service.t
+  val run : ?identifier:string -> unit -> unit Lwt.t
+  val register : ?identifier:string -> unit -> Sihl.Container.Service.t
+  val register_worker : unit -> Sihl.Container.Service.t
 end