+++ pool/routes/routes.ml
@@ -56,7 +56,11 @@ module Public = struct
               [ CustomMiddleware.Guardian.require_user_type_of
                   Pool_context.UserType.[ Guest; Contact; Admin ]
               ]
-            [ get "/index" index; get "/custom/assets/index.css" index_css ]
+            [ get "/index" index
+            ; get "/custom/assets/index.css" index_css
+            ; get "/credits" credits
+            ; get "/privacy-policy" privacy_policy
+            ]
         ; choose
             ~middlewares:
               [ CustomMiddleware.Guardian.require_user_type_of