+++ CHANGELOG.md
@@ -4,6 +4,14 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 ## [unreleased](https://github.com/uzh/pool/tree/HEAD)
 
+### Added
+
+- readonly admin comment on contacts
+
+### Changed
+
+- log failed login attempt as warning
+
 ## [0.3.3](https://github.com/uzh/pool/tree/0.3.3) - 2023-08-02
 
 ### Added