+++ CHANGELOG.md
@@ -11,6 +11,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 ### Changed
 
 - log failed login attempt as warning
+- credits and privacy policy pages
+
+### Fixed
+
+- blocking of email addresses after too many failed login attempts
 
 ## [0.3.3](https://github.com/uzh/pool/tree/0.3.3) - 2023-08-02