+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 
 - readonly admin comment on contacts
 - credits and privacy policy pages
+- basic admin statistics
 
 ### Changed